+++ Source/SwiftLintFramework/Models/PrimaryRuleList.swift
@@ -116,7 +116,7 @@ public let primaryRuleList = RuleList(rules: [
     MultipleClosuresWithTrailingClosureRule.self,
     NSLocalizedStringKeyRule.self,
     NSLocalizedStringRequireBundleRule.self,
-    NSNumberDirectInitRule.self,
+    NSNumberInitAsFunctionReferenceRule.self,
     NSObjectPreferIsEqualRule.self,
     NestingRule.self,
     NimbleOperatorRule.self,